Refactor to remove a lot of thread context switching
Make Zipper a sendable actor. Zipper's internal state is only accessed from one thread, so making it an actor means it can be sendable. The delegate causes a problem though: protocols appear to need to be nonisolated, so refactor to use a struct with closures, but closures are reference types.

Made the struct @unchecked sendable to silence the warnings. Don't think any of the data passed into a closure will be written to, or stored by multiple threads, so it's probably fine, but it still smells a bit.